+++ src/Krystal/Http/CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 1.3
 ---
 
+ * In `Krystal\Http\FileTransfer\FileEntity::getUniqueName()` forced to return file name only once
+ * In `Krystal\Http\FileTransfer\FileEntity` added `getExtension()` method that returns an extension of uploaded file
  * When returning file entities from upload, they can be used as arrays as well
  * Made file input handler more robust. Removed deep limitations. Keep indexes on initial array.
  * In `Request` service added `isLocal()` method that determines whether running from local machine or not